fix: design conflicts
