+++ web_app/hazard-reporting-map/pages/_app.tsx
@@ -2,15 +2,23 @@ import Navbar from '@/components/layouts/navbar';
 import '@/styles/globals.css';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import type { AppProps } from 'next/app';
-import { Inter } from 'next/font/google';
+import { Inter, Manrope } from 'next/font/google';
 import { Toaster } from 'sonner';
 
 const inter = Inter({ subsets: ['latin'] });
+const manrope = Manrope({
+  subsets: ['latin'],
+  weight: ['400', '600'],
+  variable: '--font-manrope',
+});
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
       <style jsx global>{`
+        :root {
+          --font-manrope: ${manrope.style.fontFamily};
+        }
         html {
           font-family: ${inter.style.fontFamily};
         }
