+++ web_app/hazard-reporting-map/styles/globals.css
@@ -2,6 +2,7 @@
 @tailwind components;
 @tailwind utilities;
 
+:root {
   --color-bg-secondary: 293, 100%, 98%; /* Background of the top bar */
   --color-bg-primary: 0, 0%, 100%; /* Background of the content area */
   --color-primary-button: 173, 75%, 41%; /* "Add button */
@@ -15,6 +16,7 @@
   --color-disable-border: 210, 36.8%, 85.1%;
   --color-success: 90, 100%, 21.6%;
   --color-purple: 258.8, 63.3%, 59.4%;
+  --color-auth-btn: 241, 28.9%, 41.4%;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -37,12 +39,9 @@
 }
 
 body {
-  /* color: rgb(var(--color-text-primary)); */
-  font-family: 'Manrope', sans-serif;
   color: black;
   background-color: hsla(var(--color-bg-primary));
   height: 100%;
-
 }
 
 @layer utilities {
@@ -58,16 +57,16 @@ body {
     @apply w-full px-4 py-2 rounded-md border border-black/20 bg-transparent;
   }
 
-  input[type="date"] {
+  input[type='date'] {
     @apply w-full px-4 py-2 rounded-md border border-black/20 bg-transparent;
   }
 
   .search-btn {
-    @apply bg-primary px-8 py-3 rounded-md block mx-auto text-white font-bold -translate-y-1/2 shadow-lg hover:shadow-primary/50 active:scale-95 transition-all;
+    @apply bg-auth-btn px-8 py-3 rounded-md block mx-auto text-white font-bold -translate-y-1/2 shadow-lg hover:shadow-primary/50 active:scale-95 transition-all;
   }
 
   .btn-primary {
-    @apply bg-primary px-8 py-2 rounded-md block text-white font-bold shadow-lg hover:shadow-primary/50 active:scale-95 transition-all;
+    @apply bg-auth-btn px-8 py-2 rounded-md block text-white font-bold shadow-lg hover:shadow-auth-btn/50 active:scale-95 transition-all;
   }
 
   .btn-disabled {
@@ -98,17 +97,17 @@ body {
     @apply flex flex-col gap-2 mb-5 w-[415px];
   }
 
-  .login-form input[type="text"],
-  .login-form input[type="password"],
-  .login-form input[type="email"] {
+  .login-form input[type='text'],
+  .login-form input[type='password'],
+  .login-form input[type='email'] {
     @apply w-[415px] h-[52px];
   }
 
   .login-form input {
-    @apply border border-black/20 rounded-md px-4 py-2;
+    @apply border border-[#666367] rounded-md px-4 py-2;
   }
 
-  input[type="checkbox"] {
+  input[type='checkbox'] {
     @apply mx-1;
   }
